Move Tampere BI implementation to core for reuse in other municipalities
