+++ service/src/integrationTest/kotlin/evaka/instance/tampere/bi/ExportBiCsvJobTest.kt
@@ -10,6 +10,9 @@ import evaka.core.application.persistence.daycare.Apply
 import evaka.core.application.persistence.daycare.CareDetails
 import evaka.core.application.persistence.daycare.Child
 import evaka.core.application.persistence.daycare.DaycareFormV0
+import evaka.core.bi.BiExportJob
+import evaka.core.bi.BiTable
+import evaka.core.bi.CSV_CHARSET
 import evaka.core.invoicing.data.upsertFeeDecisions
 import evaka.core.invoicing.data.upsertValueDecisions
 import evaka.core.invoicing.domain.ChildWithDateOfBirth
