+++ service/src/main/kotlin/evaka/core/bi/BiCsvUtils.kt
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-package evaka.instance.tampere.bi
+package evaka.core.bi
 
 import evaka.core.shared.domain.DateRange
 import evaka.core.shared.domain.TimeRange
