+++ service/src/main/kotlin/evaka/core/bi/BiExportJob.kt
@@ -2,19 +2,13 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-package evaka.instance.tampere.bi
+package evaka.core.bi
 
 import evaka.core.shared.db.Database
 import evaka.core.shared.domain.EvakaClock
-import evaka.instance.espoo.bi.CSV_CHARSET
-import evaka.instance.espoo.bi.EspooBiJob
-import evaka.instance.tampere.TampereAsyncJob
 import java.time.Duration
 
 class BiExportJob(private val client: BiExportClient) {
-    fun sendBiTable(db: Database.Connection, clock: EvakaClock, msg: TampereAsyncJob.SendBiTable) =
-        sendBiTable(db, clock, msg.table.fileName, msg.table.query)
-
     fun sendBiTable(
         db: Database.Connection,
         clock: EvakaClock,
@@ -25,7 +19,7 @@ class BiExportJob(private val client: BiExportClient) {
             tx.setStatementTimeout(Duration.ofMinutes(10))
 
             query(tx) { records ->
-                val stream = EspooBiJob.CsvInputStream(CSV_CHARSET, records)
+                val stream = CsvInputStream(CSV_CHARSET, records)
                 client.sendBiCsvFile(tableName, clock, stream)
             }
         }
