+++ service/src/main/kotlin/evaka/instance/espoo/bi/EspooBiJob.kt
@@ -4,12 +4,11 @@
 
 package evaka.instance.espoo.bi
 
+import evaka.core.bi.CSV_CHARSET
+import evaka.core.bi.CsvInputStream
 import evaka.core.shared.db.Database
 import evaka.core.shared.domain.EvakaClock
 import evaka.instance.espoo.EspooAsyncJob
-import java.io.InputStream
-import java.nio.ByteBuffer
-import java.nio.charset.Charset
 import java.time.Duration
 
 class EspooBiJob(private val client: EspooBiHttpClient) {
@@ -31,37 +30,5 @@ class EspooBiJob(private val client: EspooBiHttpClient) {
                 client.sendBiCsvFile(fileName, stream)
             }
         }
-    }
-
-    class CsvInputStream(private val charset: Charset, records: Sequence<String>) : InputStream() {
-        var totalBytes: Int = 0
-            private set
-
-        private val iterator = records.iterator()
-        private var buffer: ByteBuffer? = null
-
-        private fun acquireBuffer(): ByteBuffer? {
-            if (buffer?.hasRemaining() != true) {
-                buffer = null
-                while (iterator.hasNext()) {
-                    val bytes = iterator.next().toByteArray(charset)
-                    if (bytes.isNotEmpty()) {
-                        buffer = ByteBuffer.wrap(bytes)
-                        break
-                    }
-                }
-            }
-            return buffer
-        }
-
-        override fun read(): Int =
-            when (val buffer = acquireBuffer()) {
-                null -> -1
-
-                // end of stream
-                else -> buffer.get().toInt().also { totalBytes += 1 }
-            }
-
-        override fun available(): Int = acquireBuffer()?.remaining() ?: 0
     }
 }
